tk fix thanks to Guenter Kreidl an revert np dtypes compatible with pypy numpy

+++ pi3d/Buffer.py
@@ -66,14 +66,14 @@ class Buffer(Loadable):
     """
     #self.shape = shape
     self.textures = []
-    pts = np.array(pts, dtype=c_float)
-    texcoords = np.array(texcoords, dtype=c_float)
-    faces = np.array(faces, dtype=c_short)
+    pts = np.array(pts, dtype=float)
+    texcoords = np.array(texcoords, dtype=float)
+    faces = np.array(faces, dtype=int)
 
     if normals == None: #i.e. normals will only be generated if explictly None
       LOGGER.debug('Calculating normals ...')
 
-      normals = np.zeros(pts.shape, dtype=c_float) #empty array rights size
+      normals = np.zeros(pts.shape, dtype=float) #empty array rights size
 
       fv = pts[faces] #expand faces with x,y,z values for each vertex
       #cross product of two edges of triangles
